requirements.txt
asgiref==3.2.10
attrs==20.2.0
blessings==1.7
bpython==0.19
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
curtsies==0.3.4
Django==3.1.1
django-simple-history==2.11.0
djangorestframework==3.11.1
greenlet==0.4.16
idna==2.10
iniconfig==1.0.1
more-itertools==8.5.0
nose==1.3.7
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
Pygments==2.6.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
